turnType Class for pre-definded turns
added a class with static methods that can be called anywhere that organizes the different ways to turn in the direction dictated by the turn sequence. Now works with the more oblique turns both in and out of the turn area.

Still need to add a more comprehensive system for chosing the sequence of turns, likely to make three pre-determined sequences of pathing that correspond to the different drop off locations. These can be chosen by the initial sequence.